cpu_avatar_endpoint/requirements.txt
bithuman==0.5.22
livekit-agents[bithuman]==1.2.2
uvicorn
fastapi
dotenv
python-multipart